+++ src/aws/aws_sdk/auth.rs
@@ -2,11 +2,11 @@ use crate::aws::auth::AwsAuthentication;
 use aws_config::default_provider::credentials::default_provider;
 use aws_config::meta::credentials::LazyCachingCredentialsProvider;
 use aws_config::profile::ProfileFileCredentialsProvider;
-use aws_config::provider_config::ProviderConfig;
+
 use aws_config::sts::AssumeRoleProviderBuilder;
-use aws_types::config::{Builder, Config};
+
 use aws_types::credentials::SharedCredentialsProvider;
-use aws_types::region::Region;
+
 use aws_types::Credentials;
 
 impl AwsAuthentication {
