+++ src/aws/rusoto/auth.rs
@@ -1,12 +1,10 @@
-use rusoto_core::Region;
 use crate::aws::auth::AwsAuthentication;
 use crate::aws::rusoto::AwsCredentialsProvider;
+use rusoto_core::Region;
 
 const AWS_DEFAULT_PROFILE: &'static str = "default";
 
 impl AwsAuthentication {
-
-
     pub fn build(
         &self,
         region: &Region,
@@ -55,5 +53,35 @@ impl AwsAuthentication {
             }
             Self::Default {} => AwsCredentialsProvider::new(region, old_assume_role),
         }
+    }
+}
+#[cfg(test)]
+mod test {
+
+    #[test]
+    fn parsing_credentials_file() {
+        let tmpdir = tempfile::tempdir().unwrap();
+        let tmpfile_path = tmpdir.path().join("credentials");
+        let mut tmpfile = File::create(&tmpfile_path).unwrap();
+
+        writeln!(
+            tmpfile,
+            r#"
+            [default]
+            aws_access_key_id = default-access-key-id
+            aws_secret_access_key = default-secret
+        "#
+        )
+        .unwrap();
+
+        let auth = AwsAuthentication::File {
+            credentials_file: tmpfile_path.to_str().unwrap().to_string(),
+            profile: Some("default".to_string()),
+        };
+        let result = auth.build(&Region::AfSouth1, None).unwrap();
+        assert!(matches!(result, AwsCredentialsProvider::File { .. }));
+
+        drop(tmpfile);
+        tmpdir.close().unwrap();
     }
 }
