+++ src/sources/aws_sqs/config.rs
@@ -4,9 +4,9 @@ use crate::codecs::{DecodingConfig, FramingConfig, ParserConfig};
 use crate::config::{DataType, SourceConfig, SourceContext};
 use crate::serde::{default_decoding, default_framing_message_based};
 use crate::sources::aws_sqs::source::SqsSource;
-use aws_sdk_sqs::Endpoint;
-use aws_types::region::Region;
-use http::Uri;
+
+
+
 use serde::{Deserialize, Serialize};
 use std::cmp;
